remove old files and refactor
