--- src/dat_analysis/old/dat_object/__init__.py
@@ -1,9 +1,0 @@
-"""This Package contains my main Dat object class and all of the attributes that are associated
-with it. Should be built using a builder from DatBuilders (i.e. to build the DatHDF initially). And
-then the Dat should be initialized with a DatLoader (which shouldn't have to do much, but may be
-important for maintaining backwards compatibility)"""
-
-# import dat_analysis.dat_object.DatBuilder
-# import dat_analysis.dat_object.DatLoaders
-# import dat_analysis.dat_object.DatHDF
-# import dat_analysis.dat_object.Make_Dat
